Do not default to invalid role
**What**
- If a default role is not specified, don't set one for the user

**Why**
Snowflake allows a user to be configured with a default role, the role they take on automatically when logging in. When a default is not specified, snowflake sets it to "PUBLIC". "NULL" has no meaning unless we configured such a role and as such, is really not a good default for a terraform provider.