Ensure sql.NullString is not passed to Set
**What**
- Use correct types for values passed to `Set`

**Why**
The package that terraform uses `github.com/mitchellh/mapstructure` to decode values, does not support `sql.NullString`.

All the calls to `d.Set` were returning errors. For whatever reason, ignoring errors is common practice throughout various terraform providers so we're not changing that here.

This change passes the correct primitives and in our cases, none of the values can be null in the database so there is no need for `sql.NullString`.